It's time to deprecate Gemfile.dev.rb
Since we have the bundler.d directory, we can deprecate support for Gemfile.dev.rb
in the root of the project.  The file can move to the bundler.d directory
and we can remove some of the duplicate code found in the Gemfile.


(transferred from ManageIQ/manageiq@3a19899cef6977eeaf737517ba4b80750d283bcb)