Switch to patching to_definition instead
Technically, the patch needs to insert itself between the eval_gemfile
line and the to_definition line of Bundler::Dsl.evaluate.  The previous
attempts tried to patch eval_gemfile with a post-call, but in cases like
bundle check, the loading of the patch itself can only happen as early
as during the execution of eval_gemfile, making it very complicated to
modify that context.  Instead, if we patch to_definition with a
pre-call, we achieve the same thing but in a much simpler way.