Choose new (hopefully more stable) gems for the tests
The latest rack, which was used as the "gem with no dependencies",
introduced a new dependency, complicating the specs, so this commit
replaces them with gems that following the same pattern, but haven't
been updated in a long time.

- rack-obama is replaced with rubytest
- rack is replaced with ansi

The gem versions used in these specs follow a specific pattern in order to
test various versions and overrides. rubytest 0.7.0 depends on ansi >= 0 which
has no dependencies.  This allows us to manipulate the various versions of
ansi without introducing additional dependencies complicating the test.
ansi 1.4.2, 1.4.3, 1.5.0 are the three latest versions, and most tests "start"
with ansi 1.4.3 in the base Gemfile. omg 0.0.6 is an additional gem that also
has no dependencies.