Add ruby 3.3+/bundler 2.5+ to matrix, drop ruby 2.6/bundler 2.0-2.3
Bundler 2.4 is the last version that supports pre-ruby 3.0 (2.6/2.7).
Since ruby 3.0 can be a beast to upgrade, we might want to keep
ruby 2.7 with bundler 2.4 in the matrix  as the "bridge" for anyone
trying to upgrade to ruby 3.0.  We can choose to forcefully drop support
later on.

Note, we can still claim to try to support older rubies/bundler but we
should keep our test matrix limited to what we actively support.

Run code coverage against latest ruby and bundler.